feat: add genshin impact character builder
This commit introduces a new feature: a Genshin Impact
character builder.  It includes a loading screen,
client-side component, and page for fetching and
displaying character data.  Additionally, it adds
global styles, and updates the home page to include
a UID input and button for navigation to the
character builder.  The build process is improved
with the addition of OpenNext for Cloudflare
deployment.  Finally, the .gitignore file is updated
to exclude generated files.